# driving/api-rest/contracts/examples-openapi-v3.yaml
openapi: 3.0.1
info:
  title: Quickstart service Demo
  description: Example API definition for quickstart project
  version: 1.1.0
...
paths:
...
  /vacancies:
    post:
      tags:
        - vacancies
      summary: Method to create a vacancy
      description: This method creates a new vacancy with the given body information
      operationId: createVacancy
      requestBody:
        description: vacancyRequest
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VacancyResource'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResourceResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
  /vacancies/{id}:
    get:
      tags:
        - vacancies
      summary: Method to retrieve one vacancy
      description: This method retrieves a vacancy defined by a specific id
      operationId: getVacancy
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyResourceResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResourceResponse'
components:
  schemas:
...
    ExampleResource:
      title: ExampleResource
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        creationTime:
          type: string
          format: date-time
        identificationType:
          $ref: '#/components/schemas/IdentificationTypesRequestEnum'
        identification:
          type: string
        numberOfDaysInWeek:
          type: integer
          format: int32
    VacancyResource:
      title: VacancyResource
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
        endDate:
          type: string
        onboardingDate:
          type: string
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/VacancyCandidateResource'
    VacancyCandidateResource:
      type: object
      properties:
        candidate:
          $ref: '#/components/schemas/CandidateResource'
        createdDate:
          type: string
    CandidateResource:
      type: object
      properties:
        employeeId:
          type: string
        managedGroupId:
          type: string
        name:
          type: string
    ExampleResourceResponse:
      title: ExampleResourceResponse
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ExampleResource'
    VacancyResourceResponse:
      title: VacancyResourceResponse
      type: object
      properties:
        data:
          $ref: '#/components/schemas/VacancyResource'
...
    ErrorResource:
      title: ErrorResource
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        details:
          type: array
          items:
            type: string
    ErrorResourceResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorResource'
    IdentificationTypesRequestEnum:
      type: string
      enum:
        - DNI
        - NIE
...